+++ tencentcloud/tke/v20180525/models.go
@@ -658,6 +658,9 @@ type ClusterPublicLB struct {
 
 	// 安全策略放通单个IP或CIDR(例如: "192.168.1.0/24",默认为拒绝所有)
 	SecurityPolicies []*string `json:"SecurityPolicies,omitempty" name:"SecurityPolicies" list`
+
+	// 外网访问相关的扩展参数，格式为json
+	ExtraParam *string `json:"ExtraParam,omitempty" name:"ExtraParam"`
 }
 
 type ClusterVersion struct {